Social/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# newsCrawler/news_process.py: 1
# newsCrawler/newsspider.py: 1
# newsCrawler/scrapy_selenium/http.py: 3
Scrapy == 1.5.0

# TweetCrawler.py: 3
TwitterSearch == 1.0.2

# newsCrawler/newsspider.py: 8
beautifulsoup4 == 4.6.0

# newsCrawler/comment_classifier.py: 1
nltk == 3.3

# newsCrawler/comment_classifier.py: 2
numpy == 1.14.3

# newsCrawler/process_data.py: 9
pandas == 0.22.0

# newsCrawler/news_crawler.py: 2
python_dateutil == 2.7.3

# TweetCrawler.py: 4
# newsCrawler/news_auth.py: 2
python_dotenv == 0.8.2

# newsCrawler/news_auth.py: 3
# newsCrawler/news_crawler.py: 5
requests == 2.18.1

# newsCrawler/comment_classifier.py: 3
scikit_learn == 0.19.1

